Extruded random mock logic into a separate file
